Refactor item service to use new query string generation utility for sorting and filtering
